requirements.txt
PyQt6>=6.5.0
PyQt6-WebEngine>=6.5.0
requests>=2.31.0
pandas>=2.0.0
Pillow>=9.0.0
opencv-python>=4.8.0
numpy>=1.24.0
tqdm>=4.65.0
watchdog>=3.0.0
exifread>=2.3.2
ultralytics>=8.0.0
rembg>=2.0.0
